+++ packages/integration-tests/api/api.js
@@ -241,31 +241,6 @@ async function getExecutionLogs({ prefix, executionName }) {
  * @returns {Promise<Object>} - the POST confirmation from the API
  */
 async function addCollectionApi({ prefix, collection }) {
-  deprecate('@cumulus/integration-tests/api/api.addCollectionApi', '1.18.0', '@cumulus/integration-tests/api/collections.createCollection')
-  const response = await callCumulusApi({
-    prefix: prefix,
-    payload: {
-      httpMethod: 'POST',
-      resource: '/{proxy+}',
-      path: '/collections',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(collection)
-    }
-  });
-  return verifyCumulusApiResponse(response);
-}
-
-/**
- * Add a collection to Cumulus via the API
- *
- * @param {Object} params - params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {Object} params.collection - a collection object
- * @returns {Promise<Object>} - the POST confirmation from the API
- */
-async function addCollectionApi({ prefix, collection }) {
   deprecate('@cumulus/integration-tests/api/api.addCollectionApi', '1.18.0', '@cumulus/integration-tests/api/collections.createCollection');
   const response = await callCumulusApi({
     prefix: prefix,